requirements.txt
Django==5.0.2
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
python-dotenv==1.0.1
djangorestframework-simplejwt==5.3.1
mssql-django==1.4
pyodbc==5.0.1 